Introduce check for "out of CI credits" situation
When a Cirrus CI job started by 'cirrus-run' fails due lack of CI
credits, it would look exactly like a failure in the job itself, thus
the user would not be able to distinguish the problems.

If a 'taks' failed due to CI minute lack, the 'notifications' field of
given 'task' contains the following message:

 Failed to start an instance: FAILED_PRECONDITION: Monthly compute limit exceeded!

To detect this, receive also task data and their notifications when
waiting for a job and look for the message about CI credits.

In order to see this status from scripts this will also return a new
error code '3' if the out of CI credits situation is detected.

Additionally in order to allow quick adaptation when the above string
will be changed add '--cirrus-out-of-ci-credits-message' parameter which
can configure the string.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>